Move ConfigDescription to ScriptTransformationService
