+++ bundles/org.openhab.core.automation.module.script/src/main/java/org/openhab/core/automation/module/script/ScriptTransformationServiceFactory.java
@@ -12,14 +12,8 @@
  */
 package org.openhab.core.automation.module.script;
 
-import static org.openhab.core.automation.module.script.profile.ScriptProfileFactory.PROFILE_CONFIG_URI_PREFIX;
-
-import java.net.URI;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.Dictionary;
 import java.util.Hashtable;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
@@ -27,12 +21,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import javax.script.ScriptEngine;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.core.automation.module.script.internal.ScriptEngineFactoryHelper;
-import org.openhab.core.config.core.ConfigDescription;
-import org.openhab.core.config.core.ConfigDescriptionBuilder;
-import org.openhab.core.config.core.ConfigDescriptionProvider;
-import org.openhab.core.config.core.ConfigDescriptionRegistry;
 import org.openhab.core.transform.TransformationService;
 import org.osgi.service.component.ComponentFactory;
 import org.osgi.service.component.ComponentInstance;
@@ -49,20 +38,17 @@ import org.osgi.service.component.annotations.ReferencePolicy;
  *
  * @author Jimmy Tanagra - Initial contribution
  */
-@Component(immediate = true, service = { ScriptTransformationServiceFactory.class, ConfigDescriptionProvider.class })
+@Component(immediate = true, service = { ScriptTransformationServiceFactory.class })
 @NonNullByDefault
-public class ScriptTransformationServiceFactory implements ConfigDescriptionProvider {
-    private static final URI CONFIG_DESCRIPTION_TEMPLATE_URI = URI.create(PROFILE_CONFIG_URI_PREFIX + "SCRIPT");
+public class ScriptTransformationServiceFactory {
 
-    private final ConfigDescriptionRegistry configDescRegistry;
     private final ComponentFactory<ScriptTransformationService> scriptTransformationFactory;
 
     private final Map<String, ComponentInstance<ScriptTransformationService>> scriptTransformations = new ConcurrentHashMap<>();
 
     @Activate
-    public ScriptTransformationServiceFactory(@Reference ConfigDescriptionRegistry configDescRegistry,
+    public ScriptTransformationServiceFactory(
             @Reference(target = "(component.factory=org.openhab.core.automation.module.script.transformation.factory)") ComponentFactory<ScriptTransformationService> factory) {
-        this.configDescRegistry = configDescRegistry;
         this.scriptTransformationFactory = factory;
     }
 
@@ -113,38 +99,5 @@ public class ScriptTransformationServiceFactory implements ConfigDescriptionProv
     private void unregisterService(ComponentInstance<ScriptTransformationService> instance) {
         instance.getInstance().deactivate();
         instance.dispose();
-    }
-
-    @Override
-    public Collection<ConfigDescription> getConfigDescriptions(@Nullable Locale locale) {
-        ConfigDescription template = configDescRegistry.getConfigDescription(CONFIG_DESCRIPTION_TEMPLATE_URI, locale);
-        if (template == null) {
-            return Collections.emptyList();
-        }
-
-        return scriptTransformations.keySet().stream()
-                .map(type -> ConfigDescriptionBuilder.create(URI.create(PROFILE_CONFIG_URI_PREFIX + type.toUpperCase()))
-                        .withParameters(template.getParameters()).withParameterGroups(template.getParameterGroups())
-                        .build())
-                .toList();
-    }
-
-    @Override
-    public @Nullable ConfigDescription getConfigDescription(URI uri, @Nullable Locale locale) {
-        String uriString = uri.toString();
-        if (!uriString.startsWith(PROFILE_CONFIG_URI_PREFIX)) {
-            return null;
-        }
-        String scriptType = uriString.substring(PROFILE_CONFIG_URI_PREFIX.length());
-        if (!scriptTransformations.containsKey(scriptType.toLowerCase())) {
-            return null;
-        }
-
-        ConfigDescription template = configDescRegistry.getConfigDescription(CONFIG_DESCRIPTION_TEMPLATE_URI, locale);
-        if (template == null) {
-            return null;
-        }
-        return ConfigDescriptionBuilder.create(uri).withParameters(template.getParameters())
-                .withParameterGroups(template.getParameterGroups()).build();
     }
 }
